Use `rz_read_{le,be}XX` API in TI TMS320 code
