+++ librz/asm/arch/snes/snesdis.c
@@ -31,9 +31,9 @@ static int snesDisass(int M_flag, int X_flag, ut64 pc, RzAsmOp *op, const ut8 *b
 		if (*buf == 0x44 || *buf == 0x54) { // mvp and mvn
 			buf_asm = sdb_fmt(s_op->name, buf[1], buf[2]);
 		} else if (*buf == 0x82) { // brl
-			buf_asm = sdb_fmt(s_op->name, pc + 3 + (st16)ut8p_bw(buf + 1));
+			buf_asm = sdb_fmt(s_op->name, pc + 3 + (st16)rz_read_le16(buf + 1));
 		} else {
-			buf_asm = sdb_fmt(s_op->name, ut8p_bw(buf + 1));
+			buf_asm = sdb_fmt(s_op->name, rz_read_le16(buf + 1));
 		}
 		break;
 	case SNES_OP_32BIT:
@@ -43,14 +43,14 @@ static int snesDisass(int M_flag, int X_flag, ut64 pc, RzAsmOp *op, const ut8 *b
 		if (M_flag) {
 			buf_asm = sdb_fmt("%s #0x%02x", s_op->name, buf[1]);
 		} else {
-			buf_asm = sdb_fmt("%s #0x%04x", s_op->name, ut8p_bw(buf + 1));
+			buf_asm = sdb_fmt("%s #0x%04x", s_op->name, rz_read_le16(buf + 1));
 		}
 		break;
 	case SNES_OP_IMM_X:
 		if (X_flag) {
 			buf_asm = sdb_fmt("%s #0x%02x", s_op->name, buf[1]);
 		} else {
-			buf_asm = sdb_fmt("%s #0x%04x", s_op->name, ut8p_bw(buf + 1));
+			buf_asm = sdb_fmt("%s #0x%04x", s_op->name, rz_read_le16(buf + 1));
 		}
 		break;
 	}
