+++ librz/asm/arch/tms320/c55x_plus/c55plus.h
@@ -10,7 +10,6 @@
 #include <rz_types.h>
 #include <rz_lib.h>
 
-#include "../tms320_p.h"
 #include "../tms320_dasm.h"
 
 extern int c55x_plus_disassemble(tms320_dasm_t *dasm, const ut8 *buf, int len);
